adjusted data structure for money which is now stored as integer in firestore in order to maintain correct value

+++ lib/app_container.dart
@@ -33,7 +33,7 @@ class AppContainer {
   TransferConfirmCubit transferConfirmCubit({
     required String receiverId,
     required String receiverName,
-    required double amountToSend,
+    required int amountToSend,
   }) =>
       TransferConfirmCubit(
         receiverId: receiverId,
